+++ website/src/components/DarkModeButton.tsx
@@ -2,8 +2,16 @@ import { DarkThemeToggle } from "flowbite-react";
 import { FaSun, FaMoon } from "react-icons/fa"; // Import icons from react-icons
 import { useState, useEffect } from "react";
 
-function DarkModeButton() {
+type DarkModeButtonProps = {
+  // Optional className prop for additional styling
+  className?: string;
+};
+
+function DarkModeButton(props: DarkModeButtonProps) {
   const [isDarkMode, setIsDarkMode] = useState(false);
+
+  // Default to an empty string if no className is provided
+  const customClasses = props.className ?? "";
 
   useEffect(() => {
     const theme = localStorage.getItem("theme");
@@ -11,17 +19,18 @@ function DarkModeButton() {
     console.log("Inside useEffect");
   }, [isDarkMode]);
 
+  const classes =
+    `bg-dark-500 linear apply-fade-in-out-animation
+    rounded-2xl p-2 text-[#FFD700] transition-all
+    duration-[1s] hover:bg-[#ffffff3d]
+    focus:ring-transparent dark:text-white
+    dark:focus:ring-transparent
+    ` + customClasses;
+
   return (
     <div className="flex justify-center">
       <DarkThemeToggle
-        className="bg-dark-500 linear
-        apply-fade-in-out-animation
-          rounded-2xl p-2
-          text-[#FFD700] transition-all
-          duration-[1s] hover:bg-[#ffffff3d]
-            focus:ring-transparent dark:text-white
-            dark:focus:ring-transparent
-          "
+        className={classes}
         iconLight={FaSun}
         iconDark={FaMoon}
       />