+++ website/src/components/SearchBox.tsx
@@ -1,39 +1,85 @@
-import { IoSearch } from "react-icons/io5";
 import "../styles/Searchbox.css";
 
-type SearchBoxProps = {
-  width?: string;
-  height?: string;
-
-  borderColor?: string;
-  backgroundColor?: string;
-};
-
-const SearchBox = (props: SearchBoxProps) => {
-  // If one or more of the following props were
-  // not provided (undefined) then default to:
-  const width = props.width ?? "200px";
-  const height = props.height ?? "1.5rem";
-
+const SearchBox = () => {
   return (
     <>
-      <div id="searchbox-wrapper" className="searchbox">
-        <input
-          type="text"
-          placeholder="Search a card..."
-          style={{
-            width: width,
-            height: height,
-          }}
-          className="
-                  search-box-dark-mode-styling
-                  search-box-focus-styling
-                  rounded-lg
-                  border border-gray-500 
-                  px-4 py-2
-                  "
-        />
-        <IoSearch className="search-icon hover:cursor-pointer dark:fill-white" />
+      <div className="flex md:order-2">
+        <button
+          type="button"
+          data-collapse-toggle="navbar-search"
+          aria-controls="navbar-search"
+          aria-expanded="false"
+          className="me-1 rounded-lg p-2.5 text-sm text-gray-500 hover:bg-gray-100 focus:outline-none focus:ring-4 focus:ring-gray-200 dark:text-gray-400 dark:hover:bg-gray-700 dark:focus:ring-gray-700 md:hidden"
+        >
+          {/* SearchBox's Search Icon Responsive */}
+          <svg
+            className="size-5"
+            aria-hidden="true"
+            xmlns="http://www.w3.org/2000/svg"
+            fill="none"
+            viewBox="0 0 20 20"
+          >
+            <path
+              stroke="currentColor"
+              stroke-linecap="round"
+              stroke-linejoin="round"
+              stroke-width="2"
+              d="m19 19-4-4m0-7A7 7 0 1 1 1 8a7 7 0 0 1 14 0Z"
+            />
+          </svg>
+          <span className="sr-only">Search</span>
+        </button>
+        <div className="relative hidden md:block">
+          {/* SearchBox's Search Icon */}
+          <div className="pointer-events-none absolute inset-y-0 start-0 flex items-center ps-3">
+            <svg
+              className="size-4 text-gray-500 dark:text-gray-400"
+              aria-hidden="true"
+              xmlns="http://www.w3.org/2000/svg"
+              fill="none"
+              viewBox="0 0 20 20"
+            >
+              <path
+                stroke="currentColor"
+                stroke-linecap="round"
+                stroke-linejoin="round"
+                stroke-width="2"
+                d="m19 19-4-4m0-7A7 7 0 1 1 1 8a7 7 0 0 1 14 0Z"
+              />
+            </svg>
+            <span className="sr-only">Search icon</span>
+          </div>
+          <input
+            type="text"
+            id="search-navbar"
+            className="block w-full rounded-lg border border-gray-300 bg-gray-50 p-2 ps-10 text-sm text-gray-900 focus:border-blue-500 focus:ring-blue-500 dark:border-gray-600 dark:bg-gray-700 dark:text-white dark:placeholder:text-gray-400 dark:focus:border-blue-500 dark:focus:ring-blue-500"
+            placeholder="Search..."
+          />
+        </div>
+        <button
+          data-collapse-toggle="navbar-search"
+          type="button"
+          className="inline-flex size-10 items-center justify-center rounded-lg p-2 text-sm text-gray-500 hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-gray-200 dark:text-gray-400 dark:hover:bg-gray-700 dark:focus:ring-gray-600 md:hidden"
+          aria-controls="navbar-search"
+          aria-expanded="false"
+        >
+          <span className="sr-only">Open main menu</span>
+          <svg
+            className="size-5"
+            aria-hidden="true"
+            xmlns="http://www.w3.org/2000/svg"
+            fill="none"
+            viewBox="0 0 17 14"
+          >
+            <path
+              stroke="currentColor"
+              stroke-linecap="round"
+              stroke-linejoin="round"
+              stroke-width="2"
+              d="M1 1h15M1 7h15M1 13h15"
+            />
+          </svg>
+        </button>
       </div>
     </>
   );